refactor: location settings
