+++ apps/web/src/app/(dashboard)/(management)/locatie/[location]/instellingen/page.tsx
@@ -1,59 +1,10 @@
-import * as Headless from "@headlessui/react";
-import type React from "react";
-import type { PropsWithChildren } from "react";
 import { Divider } from "~/app/(dashboard)/_components/divider";
-import {
-  FieldGroup,
-  Fieldset,
-  Legend,
-} from "~/app/(dashboard)/_components/fieldset";
-import { Heading, Subheading } from "~/app/(dashboard)/_components/heading";
-import { Input, InputGroup } from "~/app/(dashboard)/_components/input";
-import { MediaDropzone } from "~/app/(dashboard)/_components/media-dropzone";
+import { Heading } from "~/app/(dashboard)/_components/heading";
 import { Text, TextLink } from "~/app/(dashboard)/_components/text";
-import { Textarea } from "~/app/(dashboard)/_components/textarea";
-import {
-  Facebook,
-  Google,
-  Instagram,
-  LinkedIn,
-  TikTok,
-  Whatsapp,
-  X,
-  YouTube,
-} from "~/app/_components/socials";
 import { retrieveLocationByHandle } from "~/lib/nwd";
 import LogosForm from "./_components/logos-form";
 import SettingsForm from "./_components/settings-form";
 import SocialsForm from "./_components/socials-form";
-
-function FieldSection({
-  label,
-  description,
-  children,
-}: PropsWithChildren<{
-  label: React.ReactNode;
-  description?: React.ReactNode;
-}>) {
-  return (
-    <Headless.Field
-      as="section"
-      className="gap-x-8 gap-y-6 grid sm:grid-cols-2"
-    >
-      <div className="space-y-1">
-        <Headless.Label>
-          <Subheading>{label}</Subheading>
-        </Headless.Label>
-        {description ? (
-          <Headless.Description>
-            <Text>{description}</Text>
-          </Headless.Description>
-        ) : null}
-      </div>
-      <div>{children}</div>
-    </Headless.Field>
-  );
-}
 
 export default async function Page(props: {
   params: Promise<{
@@ -75,10 +26,6 @@ export default async function Page(props: {
     socialMedia,
   } = location;
 
-  const findSocialMediaUrl = (
-    platform: (typeof socialMedia)[number]["platform"],
-  ) => socialMedia.find((media) => media.platform === platform)?.url;
-
   return (
     <div className="mx-auto max-w-4xl">
       <Heading>Instellingen</Heading>
@@ -91,273 +38,30 @@ export default async function Page(props: {
       </Text>
       <Divider className="mt-6 mb-10" />
 
-      <SettingsForm locationId={location.id}>
-        <FieldSection
-          label="Naam"
-          description="De naam zoals deze getoond wordt op de website."
-        >
-          <Input name="name" defaultValue={name ?? undefined} required />
-        </FieldSection>
-
-        <Divider soft className="my-10" />
-
-        <FieldSection
-          label="Bio"
-          description="Een korte omschrijving zoals deze getoond wordt op de website. Maximaal 260 karakters."
-        >
-          <Textarea
-            name="shortDescription"
-            defaultValue={shortDescription ?? undefined}
-            rows={3}
-            maxLength={260}
-          />
-        </FieldSection>
-
-        <Divider soft className="my-10" />
-
-        <FieldSection
-          label="E-mail"
-          description="Hoe kunnen deelnemers contact met jullie opnemen?"
-        >
-          <Input
-            name="email"
-            defaultValue={email ?? undefined}
-            type="email"
-            required
-          />
-        </FieldSection>
-
-        <Divider soft className="my-10" />
-
-        <FieldSection
-          label="Website"
-          description="Waar kunnen deelnemers meer informatie vinden?"
-        >
-          <Input
-            name="websiteUrl"
-            defaultValue={websiteUrl ?? undefined}
-            required
-            type="url"
-          />
-        </FieldSection>
-
-        <Divider soft className="my-10" />
-      </SettingsForm>
+      <SettingsForm
+        locationId={location.id}
+        name={name}
+        shortDescription={shortDescription}
+        email={email}
+        websiteUrl={websiteUrl}
+      />
 
       <Divider className="my-12" />
 
-      <LogosForm locationId={location.id}>
-        <Fieldset>
-          <Legend>
-            <Subheading className="select-none">Logo's</Subheading>
-          </Legend>
-          <Text>
-            We gebruiken verschillende variaties. Zorg dat ze allemaal
-            up-to-date zijn.
-          </Text>
-
-          <FieldGroup>
-            <Divider soft className="my-10" />
-
-            <FieldSection
-              label="Standaard"
-              description="Het algemene logo van de vaarlocatie."
-            >
-              <MediaDropzone
-                name="logo"
-                defaultPreview={
-                  logo
-                    ? {
-                        preview: logo.url,
-                        type: "image",
-                      }
-                    : undefined
-                }
-              />
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection
-              label="Icoon"
-              description="Een kleine vierkante versie van het logo, die wordt gebruik op plekken met minder ruimte. Kan zowel vierkant als rond weergegeven worden."
-            >
-              <MediaDropzone
-                name="logoSquare"
-                defaultPreview={
-                  logoSquare
-                    ? {
-                        preview: logoSquare.url,
-                        type: "image",
-                      }
-                    : undefined
-                }
-              />
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection
-              label="Diploma"
-              description="Het logo zoals hij geprint wordt op diploma's."
-            >
-              <MediaDropzone
-                name="logoCertificate"
-                defaultPreview={
-                  logoCertificate
-                    ? {
-                        preview: logoCertificate.url,
-                        type: "image",
-                      }
-                    : undefined
-                }
-              />
-            </FieldSection>
-          </FieldGroup>
-        </Fieldset>
-        <Divider soft className="my-10" />
-      </LogosForm>
+      <LogosForm
+        locationId={location.id}
+        logo={logo}
+        logoSquare={logoSquare}
+        logoCertificate={logoCertificate}
+      />
 
       <Divider className="my-12" />
 
-      <SocialsForm locationId={location.id}>
-        <Fieldset>
-          <Legend>
-            <Subheading className="select-none">Socials</Subheading>
-          </Legend>
-          <Text>Beheer de links naar verschillende social media kanalen.</Text>
-
-          <FieldGroup>
-            <Divider soft className="my-10" />
-
-            <FieldSection
-              label="Google Place ID"
-              description={
-                <span>
-                  Te vinden via{" "}
-                  <TextLink
-                    href="https://developers.google.com/maps/documentation/places/web-service/place-id#find-id"
-                    target="_blank"
-                  >
-                    developers.google.com
-                  </TextLink>
-                  . Wordt gebruikt voor het tonen van de locatie op de
-                  vaarlocatie kaart.
-                </span>
-              }
-            >
-              <InputGroup>
-                <Google data-slot="icon" />
-                <Input
-                  name="googlePlaceId"
-                  defaultValue={googlePlaceId ?? undefined}
-                  type="text"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="Facebook">
-              <InputGroup>
-                <Facebook data-slot="icon" />
-                <Input
-                  name="socials-facebook"
-                  defaultValue={findSocialMediaUrl("facebook") ?? undefined}
-                  type="url"
-                  placeholder="https://www.facebook.com/[pagina]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="Instagram">
-              <InputGroup>
-                <Instagram data-slot="icon" />
-                <Input
-                  name="socials-instagram"
-                  defaultValue={findSocialMediaUrl("instagram") ?? undefined}
-                  type="url"
-                  placeholder="https://www.instagram.com/[account]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="LinkedIn">
-              <InputGroup>
-                <LinkedIn data-slot="icon" />
-                <Input
-                  name="socials-linkedin"
-                  defaultValue={findSocialMediaUrl("linkedin") ?? undefined}
-                  type="url"
-                  placeholder="https://www.linkedin.com/company/[bedrijfsnaam]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="TikTok">
-              <InputGroup>
-                <TikTok data-slot="icon" />
-                <Input
-                  name="socials-tiktok"
-                  defaultValue={findSocialMediaUrl("tiktok") ?? undefined}
-                  type="url"
-                  placeholder="https://www.tiktok.com/@[account]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="WhatsApp">
-              <InputGroup>
-                <Whatsapp data-slot="icon" />
-                <Input
-                  name="socials-whatsapp"
-                  defaultValue={findSocialMediaUrl("whatsapp") ?? undefined}
-                  type="url"
-                  placeholder="https://wa.me/[telefoonnummer]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="X">
-              <InputGroup>
-                <X data-slot="icon" />
-                <Input
-                  name="socials-x"
-                  defaultValue={findSocialMediaUrl("x") ?? undefined}
-                  type="url"
-                  placeholder="https://x.com/[account]"
-                />
-              </InputGroup>
-            </FieldSection>
-
-            <Divider soft className="my-10" />
-
-            <FieldSection label="YouTube">
-              <InputGroup>
-                <YouTube data-slot="icon" />
-                <Input
-                  name="socials-youtube"
-                  defaultValue={findSocialMediaUrl("youtube") ?? undefined}
-                  type="url"
-                  placeholder="https://www.youtube.com/[kanaal]"
-                />
-              </InputGroup>
-            </FieldSection>
-          </FieldGroup>
-        </Fieldset>
-
-        <Divider soft className="my-10" />
-      </SocialsForm>
+      <SocialsForm
+        locationId={location.id}
+        googlePlaceId={googlePlaceId}
+        socialMedia={socialMedia}
+      />
     </div>
   );
 }
